Set gas price
Obtain gas price by simulating transaction and set it in real transaction
