+++ custodial/models/user/user_methods/methods.go
@@ -2,12 +2,11 @@
 package usermethods
 
 import (
-	"errors"
 	"fmt"
-	"strings"
 
 	"github.com/MyriadFlow/cosmos-wallet/custodial/models/user"
 	"github.com/MyriadFlow/cosmos-wallet/custodial/pkg/blockchain_cosmos"
+	"github.com/MyriadFlow/cosmos-wallet/custodial/pkg/blockchain_cosmos/cosmos_transfer"
 	"github.com/MyriadFlow/cosmos-wallet/custodial/pkg/env"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -49,13 +48,12 @@ func Transfer(uid string, from string, to string, amount int64) (txHash string, 
 		return "", fmt.Errorf("failed to get user: %w", err)
 	}
 
-	hrp := env.MustGetEnv("WALLET_ADDRESS_HRP")
-	fromAddr, err := accAddressFromBech32(from, hrp)
+	fromAddr, err := sdk.AccAddressFromBech32(from)
 	if err != nil {
 		return "", fmt.Errorf("failed to get from address: %w", err)
 	}
 
-	toAddr, err := accAddressFromBech32(to, hrp)
+	toAddr, err := sdk.AccAddressFromBech32(to)
 	if err != nil {
 		return "", fmt.Errorf("failed to get to address: %w", err)
 	}
@@ -65,9 +63,8 @@ func Transfer(uid string, from string, to string, amount int64) (txHash string, 
 	if err != nil {
 		return "", fmt.Errorf("failed to get private key: %w", err)
 	}
-
 	// Create transfer request
-	hash, err := blockchain_cosmos.Transfer(&blockchain_cosmos.TransferParams{
+	hash, err := cosmos_transfer.Transfer(&cosmos_transfer.TransacParams{
 		FromAddr: fromAddr,
 		ToAddr:   toAddr,
 		PrivKey:  privKey,
@@ -82,23 +79,4 @@ func Transfer(uid string, from string, to string, amount int64) (txHash string, 
 
 	// Return transaction hash
 	return hash, nil
-}
-
-// accAddressFromBech32 creates an AccAddress from a Bech32 string and prefix.
-func accAddressFromBech32(address string, prefix string) (addr sdk.AccAddress, err error) {
-	if len(strings.TrimSpace(address)) == 0 {
-		return sdk.AccAddress{}, errors.New("empty address string is not allowed")
-	}
-
-	bz, err := sdk.GetFromBech32(address, prefix)
-	if err != nil {
-		return nil, err
-	}
-
-	err = sdk.VerifyAddressFormat(bz)
-	if err != nil {
-		return nil, err
-	}
-
-	return sdk.AccAddress(bz), nil
 }
